the interaction router works
