+++ routers/interaction_router.py
@@ -1,0 +1,204 @@
+from mira import status, context_processor, audio_scorer, wake_word_detector, command_processor, inference_processor, logger
+from db import get_db_session
+import processors.sentence_processor as SentenceProcessor
+from models import Interaction, Person
+from fastapi import APIRouter, UploadFile, File, Form, HTTPException
+import uuid
+
+router = APIRouter(prefix="/interactions")
+
+@router.post("/register")
+async def register_interaction(audio: UploadFile = File(...), client_id: str = Form(...)):
+    """Register interaction - transcribe sentence, identify speaker, and update stream quality."""
+
+    if not status["enabled"]:
+        raise HTTPException(status_code=503, detail="Service is currently disabled")
+
+    if not client_id:
+        raise HTTPException(
+            status_code=400, detail="client_id is required for interaction registration"
+        )
+
+    sentence_buf_raw = await audio.read()
+
+    db = get_db_session()
+
+    try:
+        if len(sentence_buf_raw) == 0:
+            raise HTTPException(
+                status_code=400, detail="Received empty audio data. Please provide valid audio."
+            )
+
+        logger.info(
+            f"Processing audio data: {len(sentence_buf_raw)} bytes from client: {client_id}"
+        )
+
+        best_stream_info = audio_scorer.get_best_stream()
+        audio_float = SentenceProcessor.pcm_bytes_to_float32(sentence_buf_raw)
+        audio_scorer.update_stream_quality(client_id, audio_float)
+
+        if not best_stream_info:
+            raise HTTPException(
+                status_code=404,
+                detail="No active audio streams found. Please ensure clients are registered.",
+            )
+
+        # Check if this client has the best stream quality
+        if best_stream_info["client_id"] != client_id:
+            logger.info(
+                f"Interaction from {client_id} not registered - better stream available from {best_stream_info['client_id']} with score {best_stream_info['score']:.2f}"
+            )
+
+            return {
+                "message": "Interaction was not registered due to better audio streams",
+            }
+
+        sentence_buf = bytearray(sentence_buf_raw)
+        transcription_result = SentenceProcessor.transcribe_interaction(sentence_buf, True)
+
+        try:
+            interaction = Interaction(
+                **transcription_result,
+            )
+
+            db.add(interaction)
+            db.commit()
+            db.refresh(interaction)
+
+            logger.info(f"Interaction saved to database with ID: {interaction.id}")
+            status["recent_interactions"].append(interaction.id)
+
+            speaker = db.query(Person).filter_by(id=interaction.speaker_id).first()
+
+            if not speaker:
+                raise HTTPException(status_code=404, detail="Speaker not found")
+
+            if speaker.index == 1:  # type: ignore
+                wake_word_detection = wake_word_detector.detect_wake_words_text(
+                    client_id=client_id,
+                    transcribed_text=transcription_result["text"],
+                    audio_length=len(sentence_buf_raw) / (16000 * 2),
+                )
+
+                if wake_word_detection:
+                    logger.info(
+                        f"Wake word '{wake_word_detection.wake_word}' detected in audio from client {client_id}"
+                    )
+
+                    if wake_word_detection.callback:
+                        return None
+
+                    response = command_processor.process_command(
+                        interaction=interaction,
+                    )
+
+                    db.delete(interaction)
+                    db.commit()
+
+                    if response:
+                        return {"message": response}
+
+            return interaction
+
+        except Exception as db_error:
+            logger.error(f"Database error: {db_error}", exc_info=True)
+            db.rollback()
+            raise HTTPException(status_code=500, detail=f"Database error: {str(db_error)}")
+
+    finally:
+        db.close()
+
+
+@router.get("/{interaction_id}")
+def get_interaction(interaction_id: str):
+    """Get a specific interaction by ID."""
+    try:
+        db = get_db_session()
+        interaction = db.query(Interaction).filter_by(id=uuid.UUID(interaction_id)).first()
+
+        if not interaction:
+            raise HTTPException(status_code=404, detail="Interaction not found")
+
+        db.close()
+
+        return interaction
+    except Exception as e:
+        logger.error(f"Error fetching interaction: {e}")
+        raise HTTPException(status_code=500, detail=f"Failed to fetch interaction: {str(e)}")
+
+
+@router.post("/{interaction_id}/inference")
+def interaction_inference(interaction_id: str):
+    """Inference endpoint with database-backed context integration."""
+
+    if not status["enabled"]:
+        raise HTTPException(status_code=503, detail="Service is currently disabled")
+
+    db = get_db_session()
+
+    try:
+        interaction = db.query(Interaction).filter_by(id=uuid.UUID(interaction_id)).first()
+
+        if not interaction:
+            raise HTTPException(status_code=404, detail="Interaction not found")
+
+        # Use database-backed context processing
+        context, has_intent = context_processor.build_context(interaction)
+
+        if not has_intent:
+            return {"message": "Intent not recognized, no inference performed."}
+
+        action = inference_processor.extract_action(interaction=interaction, context=context)  # type: ignore
+
+        return action
+
+    except Exception as e:
+        logger.error(f"Error in inference: {e}")
+        raise HTTPException(status_code=500, detail=f"Inference failed: {str(e)}")
+
+    finally:
+        db.close()
+
+
+@router.get("/", deprecated=True)
+def get_interactions(limit: int = 0):
+    """Get recent interactions for live interaction display."""
+    try:
+        db = get_db_session()
+        query = db.query(Interaction).order_by(Interaction.timestamp.desc())
+        if limit != 0:
+            query = query.limit(limit)
+        interactions = query.all()
+
+        db.close()
+
+        return interactions
+    except Exception as e:
+        logger.error(f"Error fetching interactions: {e}")
+        raise HTTPException(status_code=500, detail=f"Failed to fetch interactions: {str(e)}")
+
+
+@router.delete("/{interaction_id}")
+def delete_interaction(interaction_id: str):
+    """Delete a specific interaction from the database."""
+    try:
+        db = get_db_session()
+        try:
+            interaction = db.query(Interaction).filter_by(id=uuid.UUID(interaction_id)).first()
+            if not interaction:
+                raise HTTPException(status_code=404, detail="Interaction not found")
+
+            db.delete(interaction)
+            db.commit()
+
+            status["recent_interactions"].remove(interaction.id)
+
+            return {"detail": "Interaction deleted successfully"}
+        except Exception as e:
+            db.rollback()
+            raise e
+        finally:
+            db.close()
+    except Exception as e:
+        logger.error(f"Error deleting interaction: {e}")
+        raise HTTPException(status_code=500, detail=f"Failed to delete interaction: {str(e)}")
